Fix highscore

+++ Assets/Scripts/Core/GameEvents.cs
@@ -7,7 +7,7 @@ namespace Code.Core {
         public static event ValueSetEventHandler<bool> OnPauseStatusChanged;
         public static event ValueSetEventHandler<bool> OnCutsceneStateChanged;
         public static event ValueSetEventHandler<bool> OnEndGame;
-        public static event ValueSetEventHandler<float> OnNewRecordBeaten;
+        public static event ValueSetEventHandler<double> OnNewRecordBeaten;
         #endregion
 
         #region Properties
@@ -48,7 +48,7 @@ namespace Code.Core {
         public static void Win() => OnEndGame?.Invoke(true);
         public static void Lose() => OnEndGame?.Invoke(false);
 
-        public static void BeatHighScore(float newHighScore) => OnNewRecordBeaten?.Invoke(newHighScore);
+        public static void BeatHighScore(double newHighScore) => OnNewRecordBeaten?.Invoke(newHighScore);
         #endregion
     }
 }
